--- src/managers/__init__.py
@@ -1,8 +1,0 @@
-# This is the __init__.py file for the managers package.
-# This file is used to mark the directory as a Python package and can also be used to import specific classes or functions.
-
-# Example of importing a class from a module within the package
-# from .example_manager import ExampleManager
-
-# You can also define package-level variables or functions here if needed.
-__all__ = []
