+++ src/managers/crypto_manager.py
@@ -41,7 +41,7 @@ from cryptography.hazmat.primitives.asymmetric import rsa, padding
 from cryptography.hazmat.primitives import serialization, hashes
 
 from src.utils.utils import *
-from src.logging.logging_config import *
+from src.logging.logger import *
 
 
 class CryptoManager:
@@ -73,7 +73,9 @@ class CryptoManager:
             Path to the file containing the AES key and IV.
         """
         # Set up logging
-        self.logger: logging = setup_logger("CryptoManager")
+        self.logger: logging = Logger(
+            "CryptoManager", console_level=logging.INFO, console_logging=True
+        )
 
         self.key_file = str(get_proj_root()) + key_file
         self.private_key_file = str(get_proj_root()) + private_key_file
